+++ src/components/content/document/helpers/filters.test.ts
@@ -2,7 +2,7 @@ import { applyDocumentFilters, ZUS_CREATION_DATE_URL } from "./filters";
 import { THIRD_PARTY_SOURCE_SYSTEM } from "../../resource/helpers/filters";
 import { DocumentModel } from "@/fhir/models/document";
 
-interface SyntheticDocRefProps {
+interface MockDocRefProps {
   id: string;
   title: string;
   postRainbowCategories: boolean;
@@ -12,7 +12,7 @@ interface SyntheticDocRefProps {
   binaryCreationDate?: string;
 }
 
-const createSyntheticDocRef = (props: SyntheticDocRefProps): fhir4.DocumentReference => ({
+export const createMockDocRef = (props: MockDocRefProps): fhir4.DocumentReference => ({
   id: props.id,
   resourceType: "DocumentReference",
   status: "current",
@@ -65,7 +65,7 @@ const createSyntheticDocRef = (props: SyntheticDocRefProps): fhir4.DocumentRefer
 });
 
 describe("document filters tests", () => {
-  const commonwellDoc = createSyntheticDocRef({
+  const commonwellDoc = createMockDocRef({
     id: "1",
     title: "post rainbow commonwell",
     docRefDate: "2023-04-17T12:34:56.000Z",
@@ -74,7 +74,7 @@ describe("document filters tests", () => {
     postRainbowCategories: true,
   });
 
-  const commonwelPreRainbowDoc = createSyntheticDocRef({
+  const commonwelPreRainbowDoc = createMockDocRef({
     id: "1a",
     title: "pre rainbow commonwell",
     docRefDate: "2022-04-17T12:34:56.000Z",
@@ -83,14 +83,14 @@ describe("document filters tests", () => {
     postRainbowCategories: false,
   });
 
-  const commonwellNoDateDoc = createSyntheticDocRef({
+  const commonwellNoDateDoc = createMockDocRef({
     id: "2",
     title: "no odate commonwell",
     thirdPartyOwner: "commonwell",
     postRainbowCategories: false,
   });
 
-  const surescriptsDoc = createSyntheticDocRef({
+  const surescriptsDoc = createMockDocRef({
     id: "2",
     title: "post rainbow surescript",
     docRefDate: "2023-04-17T12:34:56.000Z",
@@ -99,7 +99,7 @@ describe("document filters tests", () => {
     postRainbowCategories: true,
   });
 
-  const questDoc = createSyntheticDocRef({
+  const questDoc = createMockDocRef({
     id: "3",
     title: "post rainbow quest",
     docRefDate: "2023-04-17T12:34:56.000Z",
@@ -108,7 +108,7 @@ describe("document filters tests", () => {
     postRainbowCategories: true,
   });
 
-  const firstPartyDoc = createSyntheticDocRef({
+  const firstPartyDoc = createMockDocRef({
     id: "4",
     title: "first party",
     docRefDate: "2023-04-17T12:34:56.000Z",
@@ -116,7 +116,7 @@ describe("document filters tests", () => {
     postRainbowCategories: false,
   });
 
-  const carequalityDoc = createSyntheticDocRef({
+  const carequalityDoc = createMockDocRef({
     id: "1",
     title: "post rainbow carequality",
     docRefDate: "2023-04-17T12:34:56.000Z",