+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # ctw-component-library
 
+## 1.53.1
+
+### Patch Changes
+
+- 4ef61e93: Unit tests for whether a resource has a renderable document.
+
+## 1.53.0
+
+### Minor Changes
+
+- 832faab1: Added analytics reporting to Telemetry class
+
+## 1.52.0
+
+### Minor Changes
+
+- 0c2d0fe6: Update telemetry metrics to support Encounters & Notes monitoring
+
+## 1.51.0
+
+### Minor Changes
+
+- c5e680ad: Remove beta indicator for Encounters & Notes
+
+### Patch Changes
+
+- ff1103c0: In drawers, do not show the Source Document link when the document is not renderable.
+- 22e68219: Show all "Show Document" buttons even if unrenderable, to prevent bug where the button does not appear for renderable documents.
+
 ## 1.50.1
 
 ### Patch Changes